repair header

+++ src/fact.cpp
@@ -1,6 +1,6 @@
 #include "fact.h"
 
-int MyMath::factorial(int n) {
+long long MyMath::factorial(int n) {
     if ( n < 0 )
 	return 0;
     long long product = 1;
